fix: math error on interest

+++ programs/orbit_len/src/constants.rs
@@ -8,4 +8,4 @@ pub const BANK_SEED: &str = "bank";
 
 pub const ORBITLEN_ACCOUNT_SEED: &str = "orbitlen_account";
 
-pub const SECONDS_PER_YEAR: i128 = 31_536_000;
+pub const SECONDS_PER_YEAR: usize = 31_536_000;
